+++ ViewModel/VulnerabilityViewModel.cs
@@ -19,6 +19,7 @@ using System.IO;
 using System.Linq;
 using System.Windows;
 using System.Data.SQLite.Linq;
+using System.Linq.Expressions;
 using Vulnerator.Helper;
 using Vulnerator.Model.BusinessLogic;
 using Vulnerator.Model.DataAccess;
@@ -37,8 +38,6 @@ namespace Vulnerator.ViewModel
         public Func<double, string> Formatter { get; set; }
         public Func<double, string> DateFormatter { get; set; }
         OpenFileDialog openFileDialog;
-        private SaveFileDialog saveExcelFile;
-        private SaveFileDialog savePdfFile;
         private BackgroundWorker backgroundWorker;
         private static readonly ILog log = LogManager.GetLogger(typeof(Logger));
         public static Stopwatch stopWatch = new Stopwatch();
@@ -142,6 +141,21 @@ namespace Vulnerator.ViewModel
                 {
                     _selectedMitigationsOrCondition = value;
                     RaisePropertyChanged("SelectedMitigationsOrCondition");
+                }
+            }
+        }
+
+        private MitigationsOrCondition _editableMitigationsOrCondition { get; set; }
+
+        public MitigationsOrCondition EditableMitigationsOrCondition
+        {
+            get => _editableMitigationsOrCondition;
+            set
+            {
+                if (_editableMitigationsOrCondition != value)
+                {
+                    _editableMitigationsOrCondition = value;
+                    RaisePropertyChanged("EditableMitigationsOrCondition");
                 }
             }
         }
@@ -984,164 +998,11 @@ namespace Vulnerator.ViewModel
             }
         }
 
-        private bool ExcelReportsAreRequired()
-        {
-            bool PoamAndRarAreNeeded = bool.Parse(ConfigAlter.ReadSettingsFromDictionary("cbPoamRar"));
-            bool SummaryTabIsNeeded = bool.Parse(ConfigAlter.ReadSettingsFromDictionary("cbAssetOverview"));
-            bool DiscrepanciesTabIsNeeded = bool.Parse(ConfigAlter.ReadSettingsFromDictionary("cbDiscrepancies"));
-            bool AcasOutputTabIsNeeded = bool.Parse(ConfigAlter.ReadSettingsFromDictionary("cbAcasOutput"));
-            bool StigDetailsTabIsNeeded = bool.Parse(ConfigAlter.ReadSettingsFromDictionary("cbStigDetails"));
-            if (PoamAndRarAreNeeded || SummaryTabIsNeeded || DiscrepanciesTabIsNeeded || AcasOutputTabIsNeeded || StigDetailsTabIsNeeded)
-            { return true; }
-            else
-            { return false; }
-        }
-
-        private bool? GetExcelReportName()
-        {
-            saveExcelFile = new SaveFileDialog();
-            saveExcelFile.AddExtension = true;
-            saveExcelFile.Filter = "Excel Files (*.xlsx)|*.xlsx";
-            saveExcelFile.DefaultExt = "xlsx";
-            saveExcelFile.Title = "Save Excel Report";
-            saveExcelFile.OverwritePrompt = true;
-            saveExcelFile.CheckPathExists = true;
-            return saveExcelFile.ShowDialog();
-        }
-
-        private string CreateExcelReports()
-        {
-            log.Info("Begin creation of " + saveExcelFile.FileName);
-            fileStopWatch.Start();
-            OpenXmlReportCreator openXmlReportCreator = new OpenXmlReportCreator();
-            if (!openXmlReportCreator.CreateExcelReport(saveExcelFile.FileName).Contains("successful"))
-            {
-                log.Error("Creation of " + saveExcelFile.FileName + " failed; Elapsed time: " + fileStopWatch.Elapsed.ToString());
-                fileStopWatch.Stop();
-                fileStopWatch.Reset();
-                return "Excel report creation error; see log for details";
-            }
-            else
-            {
-                log.Info(saveExcelFile.FileName + " created successfully; Elapsed time: " + fileStopWatch.Elapsed.ToString());
-                fileStopWatch.Stop();
-                fileStopWatch.Reset();
-                return "Excel report created successfully";
-            }
-        }
-
-        private bool PdfReportIsRequired()
-        {
-            if (bool.Parse(ConfigAlter.ReadSettingsFromDictionary("cbPdfSum")))
-            { return true; }
-            else
-            { return false; }
-        }
-
-        private bool? GetPdfReportName()
-        {
-            savePdfFile = new SaveFileDialog();
-            savePdfFile.AddExtension = true;
-            savePdfFile.Filter = "PDF Files (*.pdf)|*.pdf";
-            savePdfFile.DefaultExt = "xls";
-            savePdfFile.Title = "Save PDF Report";
-            savePdfFile.OverwritePrompt = true;
-            savePdfFile.CheckPathExists = true;
-            return savePdfFile.ShowDialog();
-        }
-
-        private string CreatePdfReport()
-        {
-            log.Info("Begin creation of " + savePdfFile.FileName);
-            fileStopWatch.Start();
-            PdfReportCreator pdfReportCreator = new PdfReportCreator();
-            if (!pdfReportCreator.PdfWriter(savePdfFile.FileName.ToString(), string.Empty).Equals("Success"))
-            {
-                log.Error("Creation of " + savePdfFile.FileName + " failed; Elapsed time: " + fileStopWatch.Elapsed.ToString());
-                fileStopWatch.Stop();
-                fileStopWatch.Reset();
-                return "PDF report creation error; see log for details";
-            }
-            else
-            {
-                log.Info(savePdfFile.FileName + " created successfully; Elapsed time: " + fileStopWatch.Elapsed.ToString());
-                fileStopWatch.Stop();
-                fileStopWatch.Reset();
-                return "PDF summary created successfully";
-            }
-        }
-
         public RelayCommand ClearIngestedFilesCommand => new RelayCommand(ClearIngestedFiles);
 
         private void ClearIngestedFiles()
         { IngestedFiles.Clear(); }
 
-        public RelayCommand ExecuteExportCommand => new RelayCommand(ExecuteExport);
-
-        private void ExecuteExport()
-        {
-            backgroundWorker = new BackgroundWorker();
-            backgroundWorker.DoWork += ExecuteExportBackgroundWorker_DoWork;
-            backgroundWorker.RunWorkerAsync();
-            backgroundWorker.Dispose();
-        }
-
-        private void ExecuteExportBackgroundWorker_DoWork(object sender, DoWorkEventArgs e)
-        { 
-            try
-            {
-                GuiFeedback guiFeedback = new GuiFeedback();
-                guiFeedback.SetFields("Generating requested reports", "Visible", false);
-                Messenger.Default.Send(guiFeedback);
-                if (Properties.Settings.Default.ReportPoamRar)
-                {
-                    if ((bool)GetExcelReportName())
-                    {
-                        OpenXmlReportCreator openXmlReportCreator = new OpenXmlReportCreator();
-                        openXmlReportCreator.CreateExcelReport(saveExcelFile.FileName);
-                    }
-                }
-                guiFeedback.SetFields("Report creation complete", "Collapsed", true);
-                Messenger.Default.Send(guiFeedback);
-            }
-            catch (Exception exception)
-            {
-                log.Error("Unable to export the requested reports.");
-                log.Debug("Exception details:", exception);
-                throw exception;
-            }
-        }
-
-        public RelayCommand SetReportRequirementCommand => new RelayCommand(SetReportRequirement);
-
-        private void SetReportRequirement()
-        {
-            backgroundWorker = new BackgroundWorker();
-            backgroundWorker.DoWork += ExecuteSetReportRequirementBackgroundWorker_DoWork;
-            backgroundWorker.RunWorkerAsync();
-            backgroundWorker.Dispose();
-        }
-
-        private void ExecuteSetReportRequirementBackgroundWorker_DoWork(object sender, DoWorkEventArgs e)
-        {
-            try
-            {
-                if (!DatabaseBuilder.sqliteConnection.State.ToString().Equals("Open"))
-                { DatabaseBuilder.sqliteConnection.Open(); }
-                using (SQLiteCommand sqliteCommand = DatabaseBuilder.sqliteConnection.CreateCommand())
-                {
-                    sqliteCommand.Parameters.Add(new SQLiteParameter("Required_Report_ID", SelectedReport.Required_Report_ID));
-                    sqliteCommand.Parameters.Add(new SQLiteParameter("Is_Report_Selected", SelectedReport.Is_Report_Selected));
-                    databaseInterface.UpdateRequiredReportSelected(sqliteCommand);
-                }
-                DatabaseBuilder.sqliteConnection.Close();
-            }
-            catch (Exception exception)
-            {
-                log.Error($"Unable to update report selection criteria for {SelectedReport.Displayed_Report_Name}");
-                log.Debug("Exception details:", exception);
-            }
-        }
 
         private List<Likelihood> PopulateLikelihoodMatrix()
         { 
@@ -1316,7 +1177,6 @@ namespace Vulnerator.ViewModel
             }
         }
 
-
         private void SetInitialSqliteParameters(string sendingEntity, SQLiteCommand sqliteCommand)
         { 
             try
@@ -1367,6 +1227,28 @@ namespace Vulnerator.ViewModel
             }
         }
 
+        public RelayCommand CalculateUpdatedLikelihoodCommand => new RelayCommand(CalculateUpdatedLikelihood);
+
+        private void CalculateUpdatedLikelihood()
+        {
+            try
+            {
+                if (string.IsNullOrWhiteSpace(EditableMitigationsOrCondition.Threat_Relevance) || string.IsNullOrWhiteSpace(EditableMitigationsOrCondition.Severity_Pervasiveness))
+                { return; }
+
+                List<Likelihood> Likelihoods = PopulateLikelihoodMatrix();
+                EditableMitigationsOrCondition.Likelihood = Likelihoods.FirstOrDefault(x => x.Relevance.Equals(EditableMitigationsOrCondition.Threat_Relevance) && x.SeverityOrPervasiveness.Equals(EditableMitigationsOrCondition.Severity_Pervasiveness))?.CalculatedLikelihood;
+                if (!string.IsNullOrWhiteSpace(EditableMitigationsOrCondition.Impact))
+                { CalculateUpdatedRisk(); }
+            }
+            catch (Exception exception)
+            {
+                log.Error("Unable to calculate likelihood.");
+                log.Debug($"Exception Details: {exception}");
+                throw exception;
+            }
+        }
+
         public RelayCommand CalculateRiskCommand => new RelayCommand(CalculateRisk);
 
         private void CalculateRisk()
@@ -1378,6 +1260,26 @@ namespace Vulnerator.ViewModel
 
                 List<Risk> Risks = PopulateRiskMatrix();
                 Risk = Risks.FirstOrDefault(x => x.Likelihood.Equals(Likelihood) && x.Impact.Equals(Impact))?.CalculatedRisk;
+            }
+            catch (Exception exception)
+            {
+                log.Error("Unable to calculate risk.");
+                log.Debug($"Exception Details: {exception}");
+                throw exception;
+            }
+        }
+
+        public RelayCommand CalculateUpdatedRiskCommand => new RelayCommand(CalculateUpdatedRisk);
+
+        private void CalculateUpdatedRisk()
+        {
+            try
+            {
+                if (string.IsNullOrWhiteSpace(EditableMitigationsOrCondition.Likelihood) || string.IsNullOrWhiteSpace(EditableMitigationsOrCondition.Impact))
+                { return; }
+
+                List<Risk> Risks = PopulateRiskMatrix();
+                Risk = Risks.FirstOrDefault(x => x.Likelihood.Equals(EditableMitigationsOrCondition.Likelihood) && x.Impact.Equals(EditableMitigationsOrCondition.Impact))?.CalculatedRisk;
             }
             catch (Exception exception)
             {
@@ -1435,6 +1337,7 @@ namespace Vulnerator.ViewModel
                     sqliteTransaction.Commit();
                 }
                 ClearMitigationValues();
+                ClearVulnerabilityGroupsChecked();
                 Messenger.Default.Send(new NotificationMessage<string>("ModelUpdate", "VulnerabilityModel"), MessengerToken.ModelUpdated);
             }
             catch (Exception exception)
@@ -1454,20 +1357,71 @@ namespace Vulnerator.ViewModel
                 PredisposingConditions = string.Empty;
                 TechnicalMitigations = string.Empty;
                 ProposedMitigations = string.Empty;
-                ThreatRelevance = string.Empty;
-                SeverityPervasiveness = string.Empty;
+                ThreatRelevance = null;
+                SeverityPervasiveness = null;
                 Likelihood = string.Empty;
-                Impact = string.Empty;
+                Impact = null;
                 Risk = string.Empty;
-                ResidualRisk = string.Empty;
-                ResidualRiskAfterProposed = string.Empty;
-                MitigatedStatus = string.Empty;
+                ResidualRisk = null;
+                ResidualRiskAfterProposed = null;
+                MitigatedStatus = null;
                 EstimatedCompletionDate = string.Empty;
             }
             catch (Exception exception)
             {
                 log.Error("Unable to clear mitigation parameters.");
                 log.Debug($"Exception Details: {exception}");
+            }
+        }
+
+        public RelayCommand SetEditableMitigationOrConditionCommand => new RelayCommand(SetEditableMitigationOrCondition);
+
+        private void SetEditableMitigationOrCondition()
+        {
+            try
+            {
+                EditableMitigationsOrCondition = (MitigationsOrCondition)SelectedMitigationsOrCondition;
+                ClearVulnerabilityGroupsChecked();
+                Vulnerabilities.Select(v =>
+                {
+                    if (v.Vulnerability_ID == EditableMitigationsOrCondition.Vulnerability_ID)
+                        v.IsChecked = true; return v;
+                });
+                foreach (Group group in EditableMitigationsOrCondition.Groups)
+                {
+                    Groups.Select(g =>
+                    {
+                        if (g.Group_ID == group.Group_ID)
+                            g.IsChecked = true; return g;
+                    });
+                }
+            }
+            catch (Exception exception)
+            {
+                log.Error("Unable to clear set editable mitigation.");
+                log.Debug($"Exception Details: {exception}");
+            }
+        }
+
+        private void ClearVulnerabilityGroupsChecked()
+        {
+            try
+            {
+                Vulnerabilities.Select(v =>
+                {
+                    v.IsChecked = false;
+                    return v;
+                });
+                Groups.Select(g =>
+                {
+                    g.IsChecked = false;
+                    return g;
+                });
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+                throw;
             }
         }
     }
